move the upsertion to an activity
