+++ connectors/src/connectors/google_drive/temporal/workflows.ts
@@ -25,6 +25,7 @@ const {
   garbageCollectorFinished,
   markFolderAsVisited,
   shouldGarbageCollect,
+  upsertSharedWithMeFolder,
 } = proxyActivities<typeof activities>({
   startToCloseTimeout: "20 minutes",
 });
@@ -103,6 +104,8 @@ export async function googleDriveFullSync({
       }
     }
   });
+
+  await upsertSharedWithMeFolder(connectorId);
 
   // Temp to clean up the running workflows state
   foldersToBrowse = uniq(foldersToBrowse);
